backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
scikit-learn==1.4.0
requests==2.31.0
pyyaml==6.0.1
